Remove utilisation of Github access tokens
- Refactor `ctma new` and `cmtc install` implementations to
  not utilise the Github API & access token to retrieve
  artefacts and [release] tags
- test code needs to be refactored as it doesn't compile after
  the refactoring of the main code
- Left a couple TODO that require follow-up
