+++ cmta/src/main/scala/com/lunatech/cmt/admin/Domain.scala
@@ -18,6 +18,8 @@ import sbt.io.syntax.File
 import com.lunatech.cmt.*
 import cats.syntax.either.*
 
+import com.lunatech.cmt.Domain.InstallationSource
+
 object Domain:
 
   final case class RenumberStart(value: Int)
@@ -43,31 +45,16 @@ object Domain:
 
   final case class CourseTemplate(value: Either[CmtError, InstallationSource.GithubProject])
   object CourseTemplate:
+    val GithubTemplateRegex = "([A-Za-z0-9-_]*)".r
     val GithubProjectRegex = "([A-Za-z0-9-_]*)\\/([A-Za-z0-9-_]*)".r
     val GithubProjectWithTagRegex = "([A-Za-z0-9-_]*)\\/([A-Za-z0-9-_]*)\\/(.*)".r
     def fromString(str: String): CourseTemplate =
-      if (str.indexOf("/") > 0) {
-        str match {
-          case GithubProjectRegex(organisation, project) =>
-            CourseTemplate(Right(InstallationSource.GithubProject(organisation, project, None)))
-          case GithubProjectWithTagRegex(organisation, project, tag) =>
-            CourseTemplate(Right(InstallationSource.GithubProject(organisation, project, Some(tag))))
-          case _ => CourseTemplate(s"Invalid template name: $str".toExecuteCommandErrorMessage.asLeft)
-        }
-      } else {
-        CourseTemplate(Right(InstallationSource.GithubProject("lunatech-labs", str, None)))
+      str match {
+        case GithubTemplateRegex(template) =>
+          CourseTemplate(Right(InstallationSource.GithubProject("lunatech-labs", s"cmt-template-$template", None)))
+        case GithubProjectRegex(organisation, project) =>
+          CourseTemplate(Right(InstallationSource.GithubProject(organisation, project, None)))
+        case GithubProjectWithTagRegex(organisation, project, tag) =>
+          CourseTemplate(Right(InstallationSource.GithubProject(organisation, project, Some(tag))))
+        case _ => CourseTemplate(s"Invalid template name: $str".toExecuteCommandErrorMessage.asLeft)
       }
-
-  sealed trait InstallationSource
-
-  object InstallationSource:
-    final case class LocalDirectory(value: File) extends InstallationSource
-
-    final case class ZipFile(value: File) extends InstallationSource
-
-    final case class GithubProject(organisation: String, project: String, tag: Option[String])
-      extends InstallationSource {
-      val displayName: String =
-        if tag.isEmpty then s"$organisation/$project" else s"$organisation/$project/${tag.getOrElse("")}"
-    }
-  end InstallationSource
