+++ cmtc/src/test/scala/com/lunatech/cmt/client/cli/SetCurrentCourseSpec.scala
